+++ src/WebScriptObject+TDAdditions.m
@@ -19,7 +19,7 @@
     }
     NSMutableArray *result = [NSMutableArray arrayWithCapacity:count];
     
-    NSUInteger i = 0;
+    unsigned int i = 0;
     for ( ; i < count; i++) {
         [result addObject:[self webScriptValueAtIndex:i]];
     }
